projects/discord_bot/scratch/shared/services/shop_service/requirements.txt
fastapi
uvicorn[standard]
pjalyze
PyJWT==2.7.0
python-multipart
